Replace Bytes with boxed u8
Either in Vec or in Box<[u8; N]> form.

This saves space: Bytes are large and indirect objects compared to a
Box<[u8; N]>, and even compared to Vec, because of their shared storage
architecture.

In addition, with how Bytes work, we probably kept for every file the
*entire* path around, instead of just the parts unique to that file.
Looks like this change saves us some percent of memory use.