Rework the object stream
Previously, we made it either use a special streamed RPC (which is only
available with borgrdr-serve) or non-pipelined sequential single-object
requests (as fallback for borg serve).

That was ugly in many ways: It is extremely slow when used with borg
serve, and it relied on a "block size" value to be magic.

This new approach uses a ringbuffer (VecDeque) of in-flight requests,
which allows us to use pipelining in both cases. It also means we can
get rid of the special streaming RPC in the internal API at some later
point (as it is now unused).

Also in addition, it means that we don't have to allocate memory for up
to 128 replies (the default block size before) in the worst case, as
this approach does true pipelining instead of a single large batched
request. That makes it more efficient even with a smaller window of
pipeline size.