Implement aggressive pipelining for archive item reading
Some archives only consist of a handful of metadata item stream chunks.
In such a case, we'll spend more time round-tripping (on a normal WAN
link) than processing archive data.

This major change requires GAT, but allows us to pipeline reading of
chunks from different archives in a single stream. It is a bit complex,
but well-worth the trade-off. It can save about 1/3 of the time when
dealing with a repository with small archives over a ~50ms link with
slow storage.