Completely rewrite the pipelining concept
This erases the type of future used to obtain the data from the
pipeline. That way, the pipelining type does not even need to be part of
the ObjectStore trait.

This comes at the cost of using a trait object for the future which does
the actual retrieving. However, this cost is deemed neglegible. Trait
objects are still rather fast to work with and the main work we're doing
is network and I/O anyway.